move fileID generator infrastructure to data_release_utils
